pyproject.toml
[project]
name = "bg-remover"
version = "0.1.0"
description = "Secure local AI tool for batch background removal using rembg and Pillow"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Andrei Cristian", email = "andrei@keibisoft.com" }
]
keywords = ["ai", "background-removal", "rembg", "cli"]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "rembg>=2.0.50",
    "pillow>=10.2.0",
]

[project.optional-dependencies]
cpu = ["onnxruntime>=1.17.0"]
gpu = ["onnxruntime-gpu>=1.17.0"]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
server = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "slowapi>=0.1.9",
    "python-multipart>=0.0.9",
]

[project.scripts]
bg-remover = "bg_remover.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["input_frames*", "frames*"]

[tool.uv]
package = true
